Get rid of hidden files in DirectoryList (fixes #202)
* get rid of hidden files in DirectoryList
* added secrets to .gitignore (gdrive.token, token.pickle)
* added GDrive cache to .gitignore
* added --debug comand-line argument to show debug messages,
  it's disabled by default
* removed usage of api code by nut_gui (getting gdrive token)
* extracted gdrive-related code from api to external module
* extracted some code to the network module
