+++ server/controller/api.py
@@ -1,72 +1,28 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 import json
-from nut_impl import status
-from nut_impl import nsps
-from nut_impl import printer
-import server
-from nut_impl import config
-import time
-import nut_impl
-import requests
-import sys
-from bs4 import BeautifulSoup
-import pickle
-from googleapiclient.discovery import build
-from google_auth_oauthlib.flow import InstalledAppFlow
-from google.auth.transport.requests import Request
-import hashlib
-import traceback
 import os
+import sys
+import traceback
 
-SCOPES = ['https://www.googleapis.com/auth/drive']
+import nut_impl
+import server
 
-
-def makeRequest(method, url, hdArgs={}, start=None, end=None, accept='*/*'):
-    if start is None:
-        reqHd = {
-            'Accept': accept,
-            'Connection': None,
-            'Accept-Encoding': None,
-        }
-    else:
-        reqHd = {
-            'Accept': accept,
-            'Connection': None,
-            'Accept-Encoding': None,
-            'Range': 'bytes=%d-%d' % (start, end-1),
-        }
-
-    reqHd.update(hdArgs)
-
-    r = requests.request(
-        method,
-        url,
-        headers=reqHd,
-        verify=False,
-        stream=True,
-        timeout=15
-    )
-
-    printer.debug('%s %s %s' % (method, str(r.status_code), url))
-    printer.debug(r.request.headers)
-
-    if r.status_code == 403:
-        raise IOError('Forbidden ' + r.text)
-
-    return r
+from bs4 import BeautifulSoup
+from nut_impl import config, nsps, printer, status
+from server import gdrive, network
 
 
-def success(request, response, s):
+def _success(request, response, s):
     response.write(json.dumps({'success': True, 'result': s}))
 
-
-def error(request, response, s):
+# TODO: remove unused method
+def _error(request, response, s):
     response.write(json.dumps({'success': False, 'result': s}))
 
 
 def getScan(request, response):
-    success(request, response, nut_impl.scan())
+    _success(request, response, nut_impl.scan())
 
 
 def getSearch(request, response):
@@ -126,7 +82,7 @@ def getInfo(request, response):
         response.write(json.dumps({'success': False, 'message': str(e)}))
 
 
-def serveFile(response, path, filename=None, start=None, end=None):
+def _serveFile(response, path, filename=None, start=None, end=None):
     try:
         if start is not None:
             start = int(start)
@@ -218,6 +174,7 @@ def serveFile(response, path, filename=None, start=None, end=None):
         response.write(b'')
 
 
+# TODO: very similar to _serveFile, can be refactored
 def getDownload(request, response, start=None, end=None):
     try:
         nsp = nsps.getByTitleId(request.bits[2])
@@ -308,20 +265,20 @@ def getDownload(request, response, start=None, end=None):
         response.write(b'')
 
 
-def isWindows():
+def _isWindows():
     if "win" in sys.platform[:3].lower():
         return True
     else:
         return False
 
 
-def listDrives():
+def _listDrives():
     drives = []
     for label, _ in config.paths.mapping().items():
         drives.append(label)
-    if isWindows():
-        import string
+    if _isWindows():
         import ctypes
+        import string
         kernel32 = ctypes.windll.kernel32
         bitmask = kernel32.GetLogicalDrives()
         for letter in string.ascii_uppercase:
@@ -335,7 +292,7 @@ def listDrives():
     return drives
 
 
-def isBlocked(path):
+def _isBlocked(path):
     path = path.lower()
 
     whitelist = [
@@ -367,11 +324,11 @@ def isBlocked(path):
     return True
 
 
-def isNetworkPath(url):
+def _isNetworkPath(url):
     return url.startswith('http://') or url.startswith('https://')
 
 
-def cleanPath(path=None):
+def _cleanPath(path=None):
     if not path:
         return None
 
@@ -381,7 +338,7 @@ def cleanPath(path=None):
 
     if drive in config.paths.mapping():
         url = config.paths.mapping()[drive]
-        if isNetworkPath(url):
+        if _isNetworkPath(url):
             path = os.path.join(url, '/'.join(bits))
         else:
             path = os.path.abspath(
@@ -390,7 +347,7 @@ def cleanPath(path=None):
                     '/'.join(bits)
                 )
             )
-    elif isWindows():
+    elif _isWindows():
         path = os.path.abspath(os.path.join(drive+':/', '/'.join(bits)))
     else:
         path = os.path.abspath('/'.join(bits))
@@ -398,291 +355,19 @@ def cleanPath(path=None):
     return path
 
 
-def resolveRelativeUrl(path, parent):
+def _resolveRelativeUrl(path, parent):
     if path[0] == '/':
         if len(path) > 1:
             return path[1:]
     return path
 
 
-def isValidCache(cacheFileName, expiration=10 * 60):
-    if not os.path.isfile(cacheFileName):
-        return False
-
-    if not expiration or time.time() - os.path.getmtime(cacheFileName) < \
-            expiration:
-        return True
-    return False
-
-
-def gdriveQuery(
-    service,
-    q,
-    fields=['id', 'name', 'size', 'mimeType'],
-    expiration=10 * 60,
-    teamDriveId=None
-):
-    hashText = str(teamDriveId) + str(q) + ','.join(fields)
-    cacheFileName = 'cache/gdrive/' + hashlib.md5(
-        hashText.encode()
-    ).hexdigest()
-
-    os.makedirs('cache/gdrive/', exist_ok=True)
-
-    try:
-        if isValidCache(cacheFileName, expiration=expiration):
-            with open(cacheFileName, encoding="utf-8-sig") as f:
-                return json.loads(f.read())
-    except:
-        pass
-
-    nextToken = None
-    items = []
-
-    while True:
-        if teamDriveId:
-            results = service.files().list(
-                pageSize=100,
-                teamDriveId=teamDriveId,
-                includeItemsFromAllDrives=True,
-                corpora="teamDrive",
-                supportsTeamDrives=True,
-                q=q,
-                fields="nextPageToken, files(" + ', '.join(fields) + ")",
-                pageToken=nextToken,
-            ).execute()
-        else:
-            results = service.files().list(
-                pageSize=100,
-                q=q,
-                fields="nextPageToken, files(" + ', '.join(fields) + ")",
-                pageToken=nextToken,
-            ).execute()
-        items += results.get('files', [])
-
-        if 'nextPageToken' not in results:
-            break
-
-        nextToken = results['nextPageToken']
-
-    try:
-        with open(cacheFileName, 'w') as f:
-            json.dump(items, f)
-    except:
-        pass
-
-    return items
-
-
-def gdriveDrives(service, fields=['nextPageToken', 'drives(id, name)']):
-    cacheName = hashlib.md5((','.join(fields)).encode()).hexdigest()
-    cacheFileName = 'cache/gdrive/' + cacheName
-
-    os.makedirs('cache/gdrive/', exist_ok=True)
-
-    try:
-        if isValidCache(cacheFileName):
-            with open(cacheFileName, encoding="utf-8-sig") as f:
-                return json.loads(f.read())
-    except:
-        pass
-
-    nextToken = None
-    items = []
-
-    while True:
-        results = service.drives().list(
-            pageSize=100,
-            fields=', '.join(fields),
-            pageToken=nextToken
-        ).execute()
-        items += results.get('drives', [])
-
-        if 'nextPageToken' not in results:
-            break
-        nextToken = results['nextPageToken']
-        break
-
-    try:
-        with open(cacheFileName, 'w') as f:
-            json.dump(items, f)
-    except:
-        pass
-
-    return items
-
-
-def gdriveSearchTree(pathBits, children, id=None, roots=None):
-    if id is None:
-        for name, id in roots.items():
-            if name == pathBits[0]:
-                r = gdriveSearchTree(
-                    pathBits[1:],
-                    children[id] if id in children else [],
-                    id,
-                    roots
-                )
-                if r is not None:
-                    return r
-        return None
-
-    if len(pathBits) <= 0:
-        return id
-
-    for entry in children:
-        if entry['name'] != pathBits[0]:
-            continue
-
-        folderId = entry['id']
-
-        if len(pathBits) == 1:
-            return folderId
-
-        if folderId in children:
-            for newChildren in children[folderId]:
-                r = gdriveSearchTree(
-                    pathBits[1:],
-                    newChildren,
-                    folderId,
-                    roots
-                )
-
-                if r is not None:
-                    return r
-
-    return None
-
-
 def getTeamDriveId(service, path):
-    bits = [x for x in path.replace('\\', '/').split('/') if x]
-
-    if len(bits) == 0:
-        return None
-
-    if bits[0] == 'mydrive':
-        return None
-    else:
-        for item in gdriveDrives(service):
-            id = item['id']
-            name = item['name']
-            if name == bits[0]:
-                return id
-
-    return None
-
-
-def gdriveGetFolderId(service, path):
-    bits = [x for x in path.replace('\\', '/').split('/') if x]
-
-    if len(bits) == 0:
-        return 'root'
-
-    items = []
-
-    children = {'root': []}
-    roots = {}
-
-    rootId = None
-    teamDriveId = None
-
-    if bits[0] == 'mydrive':
-        rootId = 'root'
-    else:
-        for item in gdriveDrives(service):
-            id = item['id']
-            name = item['name']
-            if name == bits[0]:
-                rootId = id
-                teamDriveId = id
-                break
-
-    if not rootId:
-        return None
-
-    if len(bits) == 1:
-        return rootId
-
-    for item in gdriveQuery(
-        service,
-        f"'{rootId}' in parents and trashed=false and mimeType = " +
-        "'application/vnd.google-apps.folder'",
-        teamDriveId=teamDriveId
-    ):
-        roots[item['name']] = item['id']
-
-    if rootId == 'root':
-        items = gdriveQuery(
-            service,
-            "mimeType = 'application/vnd.google-apps.folder' and " +
-            "trashed=false",
-            fields=['id', 'name', 'size', 'mimeType', 'parents']
-        )
-    else:
-        items = gdriveQuery(
-            service,
-            "mimeType = 'application/vnd.google-apps.folder' and " +
-            "trashed=false",
-            fields=['id', 'name', 'size', 'mimeType', 'parents'],
-            teamDriveId=rootId
-        )
-
-    for item in items:
-        if 'parents' in item:
-            for parentId in item['parents']:
-                if parentId not in children:
-                    children[parentId] = []
-                children[parentId].append(item)
-        else:
-            children['root'].append(item)
-
-    return gdriveSearchTree(bits[1:], children, None, roots)
-
-
-def getFileInfo(service, path):
-    try:
-        bits = [x for x in path.replace('\\', '/').split('/') if x]
-        dirPath = '/'.join(bits[0:-1])
-        folderId = gdriveGetFolderId(service, dirPath)
-
-        teamDriveId = getTeamDriveId(service, path)
-
-        for item in gdriveQuery(
-            service,
-            f"'{folderId}' in parents and trashed=false and mimeType != " +
-            "'application/vnd.google-apps.folder'",
-            fields=['*'],
-            teamDriveId=teamDriveId
-        ):
-            if item['name'] == bits[-1]:
-                return item
-    except:
-        raise
-    return None
+    return gdrive.get_team_gdrive_id(service, path)
 
 
 def getGdriveToken(request, response):
-    creds = None
-
-    if os.path.exists('token.pickle'):
-        with open('token.pickle', 'rb') as token:
-            creds = pickle.load(token)
-
-    if not creds or not creds.valid:
-        if creds and creds.expired and creds.refresh_token:
-            creds.refresh(Request())
-        else:
-            flow = InstalledAppFlow.from_client_secrets_file(
-                config.getGdriveCredentialsFile(), SCOPES)
-            creds = flow.run_local_server(port=0)
-
-        with open('token.pickle', 'wb') as token:
-            pickle.dump(creds, token)
-
-        with open('gdrive.token', 'w') as token:
-            token.write(json.dumps({
-                'access_token': creds.token,
-                'refresh_token': creds.refresh_token
-            }))
+    creds = gdrive.get_token()
 
     r = {}
     r['access_token'] = creds.token
@@ -693,60 +378,6 @@ def getGdriveToken(request, response):
 
     if response is not None:
         response.write(json.dumps(r))
-
-
-def listGdriveDir(path):
-    r = {'dirs': [], 'files': []}
-
-    creds = None
-
-    if os.path.exists('token.pickle'):
-        with open('token.pickle', 'rb') as token:
-            creds = pickle.load(token)
-
-    if not creds or not creds.valid:
-        if creds and creds.expired and creds.refresh_token:
-            creds.refresh(Request())
-        else:
-            flow = InstalledAppFlow.from_client_secrets_file(
-                config.getGdriveCredentialsFile(), SCOPES)
-            creds = flow.run_local_server(port=0)
-
-        with open('token.pickle', 'wb') as token:
-            pickle.dump(creds, token)
-
-    service = build('drive', 'v3', credentials=creds)
-
-    bits = [x for x in path.replace('\\', '/').split('/') if x]
-
-    if len(bits) == 0:
-        r['dirs'].append({'name': 'mydrive'})
-        for item in gdriveDrives(service):
-            r['dirs'].append({'name': item['name']})
-    else:
-        teamDriveId = getTeamDriveId(service, path)
-        for item in gdriveQuery(
-            service,
-            "'%s' in parents and trashed=false" % gdriveGetFolderId(
-                service,
-                path
-            ),
-            teamDriveId=teamDriveId
-        ):
-            o = {'name':  item['name']}
-            if 'size' in item:
-                o['size'] = int(item['size'])
-
-            if 'kind' in item:
-                o['kind'] = item['kind']
-
-            if 'mimeType' in item and item['mimeType'] == \
-                    'application/vnd.google-apps.folder':
-                r['dirs'].append(o)
-            else:
-                r['files'].append(o)
-
-    return r
 
 
 def getDirectoryList(request, response):
@@ -761,25 +392,25 @@ def getDirectoryList(request, response):
         if virtualDir == 'gdrive':
             for i in request.bits[3:]:
                 path = os.path.join(path, i)
-            r = listGdriveDir(path)
+            r = gdrive.listGdriveDir(path)
             response.write(json.dumps(r))
             return
 
         for i in request.bits[2:]:
             path = os.path.join(path, i)
 
-        path = cleanPath(path)
+        path = _cleanPath(path)
 
         r = {'dirs': [], 'files': []}
 
         if not path:
-            for d in listDrives():
+            for d in _listDrives():
                 r['dirs'].append({'name': d})
             response.write(json.dumps(r))
             return
 
-        if isNetworkPath(path):
-            x = makeRequest('GET', path)
+        if _isNetworkPath(path):
+            x = network.makeRequest('GET', path)
             soup = BeautifulSoup(x.text, 'html.parser')
             items = soup.select('a')
 
@@ -788,12 +419,13 @@ def getDirectoryList(request, response):
 
                 if href.endswith('/'):
                     r['dirs'].append({
-                        'name': resolveRelativeUrl(href, virtualDir)
+                        'name': _resolveRelativeUrl(href, virtualDir)
                     })
-                else:
-                    r['files'].append({
-                        'name': resolveRelativeUrl(href, virtualDir)
-                    })
+                    continue
+
+                r['files'].append({
+                    'name': _resolveRelativeUrl(href, virtualDir)
+                })
 
         else:
             for name in os.listdir(path):
@@ -802,7 +434,7 @@ def getDirectoryList(request, response):
                 if os.path.isdir(abspath):
                     r['dirs'].append({'name': name})
                 elif os.path.isfile(abspath):
-                    if not isBlocked(abspath):
+                    if not _isBlocked(abspath) and not nsps.is_file_hidden(abspath):
                         r['files'].append({
                             'name': name,
                             'size': os.path.getsize(abspath),
@@ -812,68 +444,6 @@ def getDirectoryList(request, response):
         response.write(json.dumps(r))
     except:
         raise IOError('dir list access denied')
-
-
-def downloadProxyFile(url, response, start=None, end=None, headers={}):
-    bytes = 0
-
-    r = makeRequest('GET', url, start=start, end=end, hdArgs=headers)
-    size = int(r.headers.get('Content-Length'))
-
-    chunkSize = 0x100000
-
-    if size >= 10000:
-
-        for chunk in r.iter_content(chunkSize):
-            response.write(chunk)
-            bytes += len(chunk)
-
-            if not config.isRunning:
-                break
-    else:
-        response.write(r.content)
-        bytes += len(r.content)
-
-    if size != 0 and bytes != size:
-        raise ValueError(
-            f'Downloaded data is not as big as expected ({bytes}/{size})!'
-        )
-
-    return bytes
-
-
-def downloadGdriveFile(response, url, start=None, end=None):
-    creds = None
-
-    if os.path.exists('token.pickle'):
-        with open('token.pickle', 'rb') as token:
-            creds = pickle.load(token)
-
-    if not creds or not creds.valid:
-        if creds and creds.expired and creds.refresh_token:
-            creds.refresh(Request())
-        else:
-            flow = InstalledAppFlow.from_client_secrets_file(
-                config.getGdriveCredentialsFile(), SCOPES)
-            creds = flow.run_local_server(port=0)
-
-        with open('token.pickle', 'wb') as token:
-            pickle.dump(creds, token)
-
-    service = build('drive', 'v3', credentials=creds)
-
-    info = getFileInfo(service, url)
-
-    if not info:
-        return server.Response404(None, response)
-
-    return downloadProxyFile(
-        'https://www.googleapis.com/drive/v3/files/%s?alt=media' % info['id'],
-        response,
-        start,
-        end,
-        headers={'Authorization': 'Bearer ' + creds.token}
-    )
 
 
 def getFile(request, response, start=None, end=None):
@@ -887,9 +457,9 @@ def getFile(request, response, start=None, end=None):
 
         for i in request.bits[2:]:
             path = os.path.join(path, i)
-        path = cleanPath(path)
+        path = _cleanPath(path)
 
-        if isBlocked(path):
+        if _isBlocked(path):
             raise IOError('access denied')
 
         if 'Range' in request.headers:
@@ -906,12 +476,12 @@ def getFile(request, response, start=None, end=None):
             path = ''
             for i in request.bits[3:]:
                 path = os.path.join(path, i)
-            return downloadGdriveFile(response, path, start=start, end=end)
+            return gdrive.downloadGdriveFile(response, path, start=start, end=end)
 
-        elif isNetworkPath(path):
-            downloadProxyFile(path, response, start=start, end=end)
+        elif _isNetworkPath(path):
+            network.downloadProxyFile(path, response, start=start, end=end)
         else:
-            return serveFile(response, path, start=start, end=end)
+            return _serveFile(response, path, start=start, end=end)
     except:
         raise IOError('file read access denied')
 
@@ -921,7 +491,7 @@ def getFileSize(request, response):
     path = ''
     for i in request.bits[2:]:
         path = os.path.join(path, i)
-    path = cleanPath(path)
+    path = _cleanPath(path)
     try:
         t['size'] = os.path.getsize(path)
         t['mtime'] = os.path.getmtime(path)
